Running program

+++ CaptainCombat/Source/Components/Move.cs
@@ -30,10 +30,10 @@ namespace CaptainCombat.Source.Components {
         public override object getData()
         {
             var obj = new { 
-                VelocityX = (int)this.Velocity.X,
-                VelocityY = (int)this.Velocity.Y,
-                AccelerationX = (int)this.Acceleration.X,
-                AccelerationY = (int)this.Acceleration.Y,
+                VelocityX = (double)this.Velocity.X,
+                VelocityY = (double)this.Velocity.Y,
+                AccelerationX = (double)this.Acceleration.X,
+                AccelerationY = (double)this.Acceleration.Y,
                 Resistance = this.Resistance,
                 ForwardVelocity = this.ForwardVelocity,
                 RotationVelocity = this.RotationVelocity,
@@ -45,8 +45,8 @@ namespace CaptainCombat.Source.Components {
 
         public override void update(JObject json)
         {
-            this.Velocity = new Vector2((int)json.SelectToken("VelocityX"), (int)json.SelectToken("VelocityY"));
-            this.Acceleration = new Vector2((int)json.SelectToken("AccelerationX"), (int)json.SelectToken("AccelerationY"));
+            this.Velocity = new Vector2((float)json.SelectToken("VelocityX"), (float)json.SelectToken("VelocityY"));
+            this.Acceleration = new Vector2((float)json.SelectToken("AccelerationX"), (float)json.SelectToken("AccelerationY"));
             this.Resistance = (double)json.SelectToken("Resistance");
             this.ForwardVelocity = (bool)json.SelectToken("ForwardVelocity");
             this.RotationVelocity = (double)json.SelectToken("RotationVelocity");
